% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rTIMESAT}
\alias{rTIMESAT}
\alias{rTIMESAT-package}
\title{rTIMESAT}
\description{
Extract Remote Sensing Vegetation Phenology by TIMESAT Fortran library
}
\keyword{Vegetation}
\keyword{package}
\keyword{phenology}
