% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSF_main.R
\name{TSF_main}
\alias{TSF_main}
\title{TSF_main}
\usage{
TSF_main(
  y,
  qc,
  nptperyear = 23,
  jobname = "TSF",
  options = NULL,
  cache = T,
  t = NULL
)
}
\arguments{
\item{y}{The input vegetation time series}

\item{qc}{the QC flag of `y`}

\item{nptperyear}{n points per year}

\item{options}{TIMESAT options, see [write_setting()]}

\item{cache}{boolean. If false, temporal files will be removed.}

\item{t}{corresponding date of `y`}
}
\description{
TIMESAT main function
}
\seealso{
[write_input()]
}
