% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSF_process.R
\name{TSF_process}
\alias{TSF_process}
\title{TSF_process}
\usage{
TSF_process(file_set = "", ncluster = 1)
}
\arguments{
\item{file_set}{file path of TIMESAT setting file.}

\item{ncluster}{How many clusters to be used in parallel mode?}
}
\description{
TSF_process
}
