% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMESAT_process.R
\name{TIMESAT_process}
\alias{TIMESAT_process}
\title{TIMESAT_process}
\usage{
TIMESAT_process(
  d,
  nptperyear = 365,
  p_trs = 0.1,
  half_win = NULL,
  cache = FALSE,
  methods = c("SG", "AG", "DL"),
  missval = -0.1,
  wmin = 0.1,
  seasonpar = 0.2,
  iters = 2
)
}
\arguments{
\item{d}{a data.frame with the columns of `t`, `y` and `w`}
}
\description{
TIMESAT_process
}
