% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_Input.R
\name{write_Input}
\alias{write_Input}
\title{write_Input}
\usage{
write_Input(x, file = "TSM_y.txt", nptperyear = 23)
}
\arguments{
\item{x}{matrix row is time, column is site.}

\item{file}{a character string naming a file.}

\item{nptperyear}{How many points per year?}
}
\description{
Write matrix into the format of TIMESAT.
}
\examples{
library(rTIMESAT)
data("MOD13A1")
df <- MOD13A1$dt[date >= ymd("20010101") & date <= ymd("20171231")]
sitename <- "CA-NS6"
d <- df[site == sitename]
nptperyear <- 23

file_y <- sprintf("TSM_# s_y.txt", sitename)
file_w <- sprintf("TSM_# s_w.txt", sitename)

write_TSM(d$EVI/1e4 , file_y, nptperyear)
write_TSM(d$SummaryQA, file_w, nptperyear)
}
